feat(swagger): Create swagger service
- Move all swagger code onto a new SwaggerService
- Add swagger settings
