+++ package.json
@@ -40,13 +40,13 @@
   },
   "homepage": "https://github.com/romakita/ts-express-decorators#readme",
   "dependencies": {
-    "@types/swagger-schema-official": "^2.0.5",
     "colors": "^1.1.2",
     "glob": "^7.0.3",
     "reflect-metadata": "^0.1.8",
     "source-map-support": "^0.4.0",
     "ts-httpexceptions": "^2.1.1",
-    "ts-log-debug": "^2.2.0"
+    "ts-log-debug": "^2.2.0",
+    "swagger-schema-official": "^2.0.0-bab6bed"
   },
   "devDependencies": {
     "@types/chai": "^3.4.32",
@@ -56,6 +56,7 @@
     "@types/node": "^7.0.0",
     "@types/superagent": "^2.0.34",
     "@types/supertest": "^2.0.0",
+    "@types/swagger-schema-official": "^2.0.5",
     "body-parser": "^1.15.1",
     "chai": "^3.5.0",
     "compression": "^1.6.1",
@@ -71,6 +72,7 @@
     "remap-istanbul": "^0.9.1",
     "rimraf": "^2.6.1",
     "supertest": "^3.0.0",
+    "swagger-ui-express": "^2.0.0",
     "tslint": "^4.4.2",
     "typescript": "^2.0.2"
   },
