+++ src/services/controller.ts
@@ -1,7 +1,3 @@
-import {
-    BodyParameter, FormDataParameter, Parameter, Path, PathParameter, QueryParameter, Schema, Spec
-} from "@types/swagger-schema-official";
-
 import {Service} from "../decorators/service";
 import {ExpressApplication} from "./express-application";
 import Controller from "../controllers/controller";
@@ -33,9 +29,7 @@ export default class ControllerService {
      *
      * @param expressApplication
      */
-    constructor (
-        @Inject(ExpressApplication) private expressApplication: ExpressApplication
-    ) {
+    constructor(@Inject(ExpressApplication) private expressApplication: ExpressApplication) {
 
     }
 
@@ -93,6 +87,7 @@ export default class ControllerService {
 
         return this;
     }
+
     /**
      * Map all controllers collected by @Controller annotation.
      */
@@ -284,145 +279,10 @@ export default class ControllerService {
         return routes;
     }
 
-
-    public getOpenAPISpec(): Spec {
-
-        let openAPI: Spec = {
-            "swagger": "2.0",
-            "info": {
-                "version": "1.0.0",
-                "title": "Swagger",
-            },
-            "paths": {}
-        };
-
-        let paths: { [pathName: string]: Path } = {};
-
-
-        const getOpenAPIPath = (expressPath: string | RegExp): string | RegExp => {
-            if (typeof expressPath === "string") {
-                let params = expressPath.match(/:[\w]+/g);
-
-                let OpenAPIPath = expressPath;
-                if (params) {
-                    let swagerParams = params.map(x => {
-                        return "{" + x.replace(":", "") + "}";
-                    });
-
-                    OpenAPIPath = params.reduce((acc, el, ix) => {
-                        return acc.replace(el, swagerParams[ix]);
-                    }, expressPath);
-                }
-                return OpenAPIPath;
-            } else {
-                return expressPath;
-            }
-        };
-
-        const getOpenAPIParams = (tsDecoratorsParams: any): Parameter[] => {
-            let OpenAPIParameters: Parameter[] = [];
-
-            let bodyParams: BodyParameter = {in: "", name: ""};
-
-            let notBodyParams: PathParameter | QueryParameter | FormDataParameter;
-            notBodyParams = {type: "", name: "", required: false, in: ""};
-
-            let hasBodyParams = false;
-            tsDecoratorsParams.forEach(p => {
-                switch (p.name) {
-                    case "parseParams":
-                        notBodyParams.in = "path";
-                        notBodyParams.required = true;
-                        notBodyParams.type = p.use ? p.use.name.toLowerCase() : "";
-                        break;
-                    case "parseBody":
-                        hasBodyParams = true;
-                        if (!bodyParams.schema) {
-                            bodyParams.schema = {};
-                            bodyParams.schema.properties = {};
-                        }
-                        let property: Schema = {
-                            type: p.use ? p.use.name.toLowerCase() : ""
-                        };
-                        bodyParams.schema.properties[p.expression] = property;
-                        if (p.required) {
-                            if (!bodyParams.schema.required) bodyParams.schema.required = [];
-                            bodyParams.schema.required.push(p.expression);
-                        }
-                        break;
-                    default:
-                        notBodyParams["in"] = "path";
-                        notBodyParams["type"] = p.use ? p.use.name.toLowerCase() : "";
-                }
-                notBodyParams.name = p.expression;
-                if (p.name !== "parseBody" && p.expression) {
-                    if (p.required) notBodyParams["required"] = true;
-                    OpenAPIParameters.push(notBodyParams);
-                }
-            });
-
-            if (hasBodyParams) {
-                bodyParams.in = "body";
-                bodyParams.name = "bodyParam";
-                bodyParams.schema.type = "object";
-                OpenAPIParameters.push(bodyParams);
-            }
-            return OpenAPIParameters;
-        };
-
-        const buildRoutes = (ctrl: Controller, endpointUrl: string) => {
-
-            ctrl.dependencies.forEach((ctrl: Controller) => buildRoutes(ctrl, `${endpointUrl}${ctrl.endpointUrl}`));
-
-            ctrl.endpoints.forEach((endpoint: Endpoint) => {
-
-                if (endpoint.hasMethod()) {
-
-                    const className = getClassName(ctrl.targetClass),
-                        parameters = InjectParams.getParams(ctrl.targetClass, endpoint.methodClassName),
-                        returnContentType = Metadata.getReturnContentType(ctrl.targetClass, endpoint.methodClassName);
-
-
-                    let OpenAPIPath = `${endpointUrl}${getOpenAPIPath(endpoint.getRoute()) || ""}`;
-
-                    if (!paths[OpenAPIPath]) paths[OpenAPIPath] = {};
-
-                    paths[OpenAPIPath][endpoint.getMethod()] = {
-                        operationId: endpoint.methodClassName,
-                        tags: [className],
-                        parameters: getOpenAPIParams(parameters),
-                        consumes: [],
-                        responses: {"200": {description: ""}},
-                        produces: returnContentType ? [returnContentType] : []
-                    };
-
-                }
-            });
-
-        };
-
-        ControllerService
-            .controllers
-            .forEach((finalCtrl: Controller) => {
-
-                if (!finalCtrl.parent) {
-                    finalCtrl
-                        .getMountEndpoints()
-                        .map(endpoint => finalCtrl.getEndpointUrl(endpoint))
-                        .forEach(endpoint => buildRoutes(finalCtrl, endpoint));
-                }
-
-
-            });
-
-        openAPI.paths = paths;
-        return openAPI;
-    }
-
     /**
      * Print all route mounted in express via Annotation.
      */
-    public printRoutes(logger: {info: (s) => void} = $log): void {
+    public printRoutes(logger: { info: (s) => void } = $log): void {
 
         const mapColor = {
             GET: (<any>"GET").green,
@@ -455,5 +315,14 @@ export default class ControllerService {
 
         logger.info("\n" + str.trim());
 
+    }
+
+    /**
+     *
+     * @param callback
+     * @param thisArg
+     */
+    public forEach(callback, thisArg?) {
+        return ControllerService.controllers.forEach(callback, thisArg);
     }
 }
