+++ src/services/route.ts
@@ -1,4 +1,3 @@
-import {Spec} from "@types/swagger-schema-official";
 import {Service} from "../decorators/service";
 import {IControllerRoute} from "../interfaces/ControllerRoute";
 import {$log} from "ts-log-debug";
@@ -20,14 +19,6 @@ export default class RouteService {
      */
     getAll(): IControllerRoute[] {
         return this.controllerService.getRoutes();
-    }
-
-    /**
-     * Return openAPISpec for routes stored in Controller manager.
-     * @returns {Spec}
-     */
-    getOpenAPISpec(): Spec {
-        return this.controllerService.getOpenAPISpec();
     }
 
     /**
